+++ rulewerk-core/src/test/java/org/semanticweb/rulewerk/core/reasoner/implementation/FileDataSourceTestUtils.java
@@ -14,9 +14,9 @@ import java.io.File;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -92,9 +92,9 @@ public final class FileDataSourceTestUtils {
 		return content;
 	}
 
-	public static void testConstructor(final FileDataSource fileDataSource, final File expectedFile,
+	public static void testConstructor(final FileDataSource fileDataSource, final String expectedFileName,
 			final String expectedDirCanonicalPath, final String expectedFileNameWithoutExtension) throws IOException {
-		assertEquals(expectedFile, fileDataSource.getFile());
+		assertEquals(expectedFileName, fileDataSource.getName());
 		assertEquals(expectedDirCanonicalPath, fileDataSource.getDirCanonicalPath());
 		assertEquals(expectedFileNameWithoutExtension, fileDataSource.getFileNameWithoutExtension());
 	}