+++ rulewerk-core/src/test/java/org/semanticweb/rulewerk/core/reasoner/implementation/VLogReasonerStateTest.java
@@ -9,9 +9,9 @@ package org.semanticweb.rulewerk.core.reasoner.implementation;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -24,7 +24,6 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
-import java.io.File;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -221,7 +220,7 @@ public class VLogReasonerStateTest {
 		// assert r(d)
 		final Predicate predicateR1 = Expressions.makePredicate("r", 1);
 		kb.addStatement(new DataSourceDeclarationImpl(predicateR1,
-				new CsvFileDataSource(new File(FileDataSourceTestUtils.INPUT_FOLDER, "constantD.csv"))));
+				new CsvFileDataSource(FileDataSourceTestUtils.INPUT_FOLDER + "constantD.csv")));
 		// p(?x) -> q(?x)
 
 		try (final VLogReasoner reasoner = new VLogReasoner(kb)) {