+++ rulewerk-parser/src/main/java/org/semanticweb/rulewerk/parser/datasources/CsvFileDataSourceDeclarationHandler.java
@@ -9,9 +9,9 @@ package org.semanticweb.rulewerk.parser.datasources;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -20,7 +20,6 @@ package org.semanticweb.rulewerk.parser.datasources;
  * #L%
  */
 
-import java.io.File;
 import java.io.IOException;
 import java.util.List;
 
@@ -42,12 +41,12 @@ public class CsvFileDataSourceDeclarationHandler implements DataSourceDeclaratio
 	public DataSource handleDirective(List<DirectiveArgument> arguments, final SubParserFactory subParserFactory)
 			throws ParsingException {
 		DirectiveHandler.validateNumberOfArguments(arguments, 1);
-		File file = DirectiveHandler.validateFilenameArgument(arguments.get(0), "source file");
+		String fileName = DirectiveHandler.validateStringArgument(arguments.get(0), "source file");
 
 		try {
-			return new CsvFileDataSource(file);
+			return new CsvFileDataSource(fileName);
 		} catch (IOException e) {
-			throw new ParsingException("Could not use source file \"" + file.getName() + "\": " + e.getMessage(), e);
+			throw new ParsingException("Could not use source file \"" + fileName + "\": " + e.getMessage(), e);
 		}
 	}
 }